WIP - Extract some methods from truffle-contract/lib/contract
