+++ packages/truffle-contract/lib/contract/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  bootstrap: require("./bootstrap"),
+  contructorMethods: require("./constructorMethods"),
+  properties: require("./properties")
+};
